# crates/synapse-protocol/schema/openapi.json
{
  "components": {
    "schemas": {
      "Conflict": {
        "additionalProperties": false,
        "properties": {
          "base_ciphertext_hash": {
            "description": "64 zeros when base_revision is 0 (no base ciphertext exists at genesis)",
            "pattern": "^[a-f0-9]{64}$"
          },
          "base_revision": {
            "description": "0 means vault genesis: the conflicting operation was built offline before any pull, so no base revision exists",
            "minimum": 0
          },
          "local_ciphertext_hash": {
            "pattern": "^[a-f0-9]{64}$"
          },
          "note_id": {
            "format": "uuid"
          },
          "operation_id": {
            "format": "uuid"
          },
          "protocol_version": {
            "const": 1
          },
          "remote_ciphertext_hash": {
            "pattern": "^[a-f0-9]{64}$"
          },
          "remote_revision": {
            "minimum": 1
          },
          "vault_id": {
            "format": "uuid"
          }
        },
        "required": [
          "protocol_version",
          "operation_id",
          "vault_id",
          "note_id",
          "base_revision",
          "remote_revision",
          "base_ciphertext_hash",
          "local_ciphertext_hash",
          "remote_ciphertext_hash"
        ],
        "type": "object"
      },
      "EncryptedPushOperation": {
        "additionalProperties": false,
        "properties": {
          "aad_version": {
            "minimum": 1,
            "type": "integer"
          },
          "base_revision": {
            "minimum": 0,
            "type": "integer"
          },
          "ciphertext": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "ciphertext_hash": {
            "pattern": "^[a-f0-9]{64}$",
            "type": "string"
          },
          "encrypted_vault_key_envelope": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "nonce": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "maxItems": 24,
            "minItems": 24,
            "type": "array"
          },
          "note_id": {
            "format": "uuid",
            "type": "string"
          },
          "operation_id": {
            "format": "uuid",
            "type": "string"
          },
          "protocol_version": {
            "const": 1,
            "type": "integer"
          },
          "vault_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "protocol_version",
          "operation_id",
          "vault_id",
          "note_id",
          "base_revision",
          "ciphertext",
          "nonce",
          "aad_version",
          "ciphertext_hash"
        ],
        "type": "object"
      },
      "PullRequest": {
        "additionalProperties": false,
        "properties": {
          "cursor": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "limit": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer"
          },
          "protocol_version": {
            "const": 1,
            "type": "integer"
          },
          "vault_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "protocol_version",
          "vault_id",
          "cursor",
          "limit"
        ],
        "type": "object"
      },
      "PullResponse": {
        "additionalProperties": false,
        "properties": {
          "next_cursor": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/EncryptedPushOperation"
            },
            "type": "array"
          },
          "protocol_version": {
            "const": 1,
            "type": "integer"
          }
        },
        "required": [
          "protocol_version",
          "operations",
          "next_cursor"
        ],
        "type": "object"
      },
      "ResnapshotRequired": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "sync_cursor_resnapshot_required",
            "type": "string"
          },
          "protocol_version": {
            "const": 1,
            "type": "integer"
          },
          "resnapshot_cursor": {
            "type": "null"
          }
        },
        "required": [
          "protocol_version",
          "code",
          "resnapshot_cursor"
        ],
        "type": "object"
      },
      "SessionResponse": {
        "additionalProperties": false,
        "properties": {
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "type": "object"
      },
      "VaultKeyEnvelope": {
        "additionalProperties": false,
        "properties": {
          "bytes": {
            "items": {
              "maximum": 255,
              "minimum": 0,
              "type": "integer"
            },
            "maxItems": 65536,
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "bytes"
        ],
        "type": "object"
      },
      "VaultListResponse": {
        "additionalProperties": false,
        "properties": {
          "vaults": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "id": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "required": [
                "id"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "vaults"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "title": "Synapse encrypted sync API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/session": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Authenticated session"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        },
        "summary": "Return the opaque authenticated user id for the current session cookie"
      }
    },
    "/v1/vaults": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultListResponse"
                }
              }
            },
            "description": "Owned or membership vault ids only"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        },
        "summary": "List opaque vault ids the authenticated user may access"
      }
    },
    "/v1/vaults/{vault_id}/envelope": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaultKeyEnvelope"
                }
              }
            },
            "description": "Opaque envelope bytes"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "404": {
            "description": "Envelope missing or vault not authorized"
          }
        },
        "summary": "Read the opaque encrypted vault-key envelope for the authenticated user"
      },
      "put": {
        "responses": {
          "204": {
            "description": "Envelope stored"
          },
          "400": {
            "description": "Invalid envelope payload"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "CSRF origin rejected"
          },
          "404": {
            "description": "Vault not authorized"
          }
        },
        "summary": "Upsert the opaque encrypted vault-key envelope for the authenticated user"
      }
    },
    "/v1/vaults/{vault_id}/operations": {
      "get": {
        "parameters": [
          {
            "description": "Opaque sync cursor; omit to resnapshot from revision 0",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": true,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponse"
                }
              }
            },
            "description": "A stable page of encrypted operations"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResnapshotRequired"
                }
              }
            },
            "description": "The opaque cursor is unknown, belongs to another vault or user, or predates retention; retry with cursor omitted"
          }
        },
        "summary": "Pull encrypted operations in strictly increasing server revision order"
      },
      "post": {
        "summary": "Push an encrypted operation"
      }
    }
  }
}